Velocity, Tempo, and Color Schemes
Redid the color scheme to make everything more uniform.  Used red for the red LEDs.  Tempo and velocity can now be changed with the buttons and the slider.  At the moment, the big button lets the user set the tempo.  I moved the code for the onClick and onSliderValueChanged to where the buttons and slider are set up instead of having the code in the overridden functions.